# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Sia - Agent Orchestration API",
    "description": "Sia connects with code generation agents and receives your requests from Slack, Discord, or through Sia mobile app. It picks up the task, schedules it, and gets it done like your personal dev assistant.",
    "version": "1.0.0"
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "additionalProperties": false,
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        },
        "required": ["error"]
      },
...
      "RepoConfig": {
        "additionalProperties": false,
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "repoId": {
            "type": "string"
          },
          "orgId": {
            "type": "string"
          },
          "executionStrategy": {
            "anyOf": [
              {
                "type": "string",
                "enum": ["auto"]
              },
              {
                "type": "string",
                "enum": ["devcontainer"]
              },
              {
                "type": "string",
                "enum": ["docker-compose"]
              },
              {
                "type": "string",
                "enum": ["custom"]
              }
            ]
          },
          "setupCommands": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "buildCommands": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "testCommands": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "validationStrategy": {
            "type": "object",
            "properties": {
              "runBuild": {
                "type": "boolean"
              },
              "runTests": {
                "type": "boolean"
              },
              "runLinter": {
                "type": "boolean"
              }
            },
            "required": ["runBuild", "runTests", "runLinter"]
          },
          "envVarsNeeded": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "detectedLanguage": {
            "type": "string"
          },
          "detectedFrom": {
            "type": "string"
          },
          "devcontainerConfig": {
            "type": "object",
            "properties": {
              "image": {
                "anyOf": [
                  {
                    "type": "string"
                  },
                  {
                    "type": "null"
                  }
                ]
              },
              "runArgs": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            },
            "required": ["image", "runArgs"]
          },
          "isConfirmed": {
            "type": "boolean"
          },
          "inferredAt": {
            "format": "date-time",
            "type": "string"
          },
          "confirmedAt": {
            "format": "date-time",
            "type": "string"
          },
          "inferenceSource": {
            "type": "string"
          },
          "inferenceConfidence": {
            "anyOf": [
              {
                "type": "string",
                "enum": ["high"]
              },
              {
                "type": "string",
                "enum": ["medium"]
              },
              {
                "type": "string",
                "enum": ["low"]
              }
            ]
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "repoId",
          "orgId",
          "isConfirmed",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "UpdateRepoDescriptionRequest": {
        "additionalProperties": false,
        "type": "object",
        "properties": {
          "description": {
            "type": "string"
          }
        }
      },
      "ConfirmRepoConfigRequest": {
        "additionalProperties": false,
        "type": "object",
        "properties": {
          "setupCommands": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "buildCommands": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "testCommands": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "GetReposResponse": {
        "type": "array",
        "items": {
          "additionalProperties": false,
          "type": "object",
          "properties": {
            "id": {
              "type": "string"
            },
            "name": {
              "type": "string"
            },
            "description": {
              "type": "string"
            },
            "url": {
              "type": "string"
            },
            "repo_provider_id": {
              "type": "string"
            },
            "created_at": {
              "format": "date-time",
              "type": "string"
            },
            "updated_at": {
              "format": "date-time",
              "type": "string"
            }
          },
          "required": [
            "id",
            "name",
            "url",
            "repo_provider_id",
            "created_at",
            "updated_at"
          ]
        }
      },
...
    }
  },
  "paths": {
...
    "/repos/github/providers/{providerId}/repos": {
      "get": {
        "tags": ["repos"],
        "description": "Get all repos for a GitHub repo provider (fetched from GitHub API)",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "providerId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "List of repos from GitHub",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetReposResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Repo provider not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/repos/{repoId}/config": {
      "get": {
        "tags": ["repos"],
        "description": "Get repository configuration",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "repoId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Repository configuration",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RepoConfig"
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "repoId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      },
      "delete": {
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "repoId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/repos/org/{orgId}/configs": {
      "get": {
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "orgId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/repos/{repoId}/config/confirm": {
      "post": {
        "tags": ["repos"],
        "description": "Confirm repository configuration",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmRepoConfigRequest"
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "repoId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Configuration confirmed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RepoConfig"
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/repos/{repoId}/config/re-infer": {
      "post": {
        "tags": ["repos"],
        "description": "Re-infer repository configuration",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "repoId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Configuration re-inferred",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RepoConfig"
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/repos/": {
      "get": {
        "tags": ["repos"],
        "description": "Get all repositories with configurations",
        "responses": {
          "200": {
            "description": "List of repositories with configurations",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetReposWithConfigsResponse"
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/repos/{id}": {
      "patch": {
        "tags": ["repos"],
        "description": "Update repository description",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRepoDescriptionRequest"
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Repository updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RepoWithConfig"
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Development server"
    }
  ],
...
}